Add 'allow_nil' option support to 'represents'
