+++ granite.gemspec
@@ -8,7 +8,7 @@ Gem::Specification.new do |s|
   s.license     = 'MIT'
 
   s.add_runtime_dependency 'actionpack', '~> 5.1'
-  s.add_runtime_dependency 'active_data', '1.1.1'
+  s.add_runtime_dependency 'active_data', '1.1.2'
   s.add_runtime_dependency 'activesupport', '~>5.1'
   s.add_runtime_dependency 'memoist', '~> 0.16'
