+++ lib/granite.rb
@@ -17,8 +17,11 @@ module Granite
   singleton_class.delegate(*Granite::Context.delegated, to: :context)
 end
 
+require 'granite/base'
 require 'granite/dispatcher'
 require 'granite/action'
 require 'granite/projector'
 require 'granite/routing'
 require 'granite/rails' if defined?(::Rails)
+
+ActiveData.base_concern = Granite::Base
