+++ spec/lib/granite/projector/helpers_spec.rb
@@ -26,6 +26,15 @@ RSpec.describe Granite::Projector::Helpers, type: :granite_projector do
     specify { expect(Granite.with_view_context(view_context) { projector.view_context }).to eq(view_context) }
   end
 
+  describe '#translate' do
+    projector { DummyAction.dummy }
+
+    specify do
+      expect(controller.translate(:key)).to eq 'translation missing: en.key'
+      expect(controller.view_context.translate(:key)).to eq '<span class="translation_missing" title="translation missing: en.key">Key</span>'
+    end
+  end
+
   context do
     projector { DummyAction.dummy }
